Promjena na hr.foi.air.cmykui i priprema za upload na jfrog
