+++ app/src/main/java/com/cmykui/framework/cmykui/MatijaPlayground.java
@@ -3,7 +3,7 @@ package com.cmykui.framework.cmykui;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-import com.cmykui.framework.cmykui.graph.BarChart;
+import hr.foi.air.cmykui.graph.BarChart;
 
 public class MatijaPlayground extends AppCompatActivity {
