+++ build.gradle
@@ -5,10 +5,14 @@ buildscript {
     repositories {
         google()
         jcenter()
+        maven {
+            url  "https://cmykui.bintray.com/maven"
+        }
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.2.1'
-
+        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.7.3'
+        classpath 'com.github.dcendents:android-maven-gradle-plugin:1.4.1'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -19,6 +23,9 @@ allprojects {
     repositories {
         google()
         jcenter()
+        maven {
+            url  "https://cmykui.bintray.com/maven"
+        }
     }
 }
