+++ cmykui/src/main/AndroidManifest.xml
@@ -1,2 +1,4 @@
+<!--<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.cmykui.framework.cmykui" />-->
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.cmykui.framework.cmykui" />
+    package="hr.foi.air.cmykui" />
