--- cmykui/src/main/java/com/cmykui/framework/cmykui/ANDROIDSTUDIOJEGLUPIOVAKLASASLUZIKAKOBIVIDIOFOLDERE.java
@@ -1,4 +1,0 @@
-package com.cmykui.framework.cmykui;
-
-public final class ANDROIDSTUDIOJEGLUPIOVAKLASASLUZIKAKOBIVIDIOFOLDERE {
-}
