--- cmykui/src/main/java/com/cmykui/framework/cmykui/base/ComponentInterface.java
@@ -1,4 +1,0 @@
-package com.cmykui.framework.cmykui.base;
-
-public interface ComponentInterface {
-}
