+++ cmykui/src/main/java/hr/foi/air/cmykui/base/ComponentInterface.java
@@ -1,0 +1,4 @@
+package hr.foi.air.cmykui.base;
+
+public interface ComponentInterface {
+}
