+++ cmykui/src/main/java/hr/foi/air/cmykui/base/OnClick.java
@@ -1,4 +1,4 @@
-package com.cmykui.framework.cmykui.base;
+package hr.foi.air.cmykui.base;
 
 public interface OnClick<V> {
     V onClick() throws Exception;
