+++ cmykui/src/main/java/hr/foi/air/cmykui/component/DialogComponent.java
@@ -1,4 +1,4 @@
-package com.cmykui.framework.cmykui.component;
+package hr.foi.air.cmykui.component;
 
 import android.app.AlertDialog;
 import android.content.Context;
