+++ cmykui/src/main/java/hr/foi/air/cmykui/graph/LineChart.java
@@ -1,4 +1,4 @@
-package com.cmykui.framework.cmykui.graph;
+package hr.foi.air.cmykui.graph;
 
 import android.content.Context;
 import android.graphics.Canvas;
