+++ cmykui/src/main/java/hr/foi/air/cmykui/layout/ImageLayout.java
@@ -1,4 +1,4 @@
-package com.cmykui.framework.cmykui.layout;
+package hr.foi.air.cmykui.layout;
 
 import android.content.Context;
 import android.graphics.Point;
@@ -10,30 +10,31 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.widget.EditText;
+import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.cmykui.framework.cmykui.R;
+import hr.foi.air.cmykui.R;
 
-public class InboxLayout extends ViewGroup {
+public class ImageLayout extends ViewGroup {
 
     int deviceWidth;
     int deviceHeight;
 
     TextView TitleText;
-    TextView SearchText;
+    ImageView SearchText;
     EditText SearchInput;
 
     private boolean SearchON = false;
     private float scale = getResources().getDisplayMetrics().density;
 
-    public InboxLayout(Context context) {
+    public ImageLayout(Context context) {
         super(context);
     }
 
-    public InboxLayout(Context context, AttributeSet attrs) {
+    public ImageLayout(Context context, AttributeSet attrs) {
         this(context, attrs, 0);
     }
-    public InboxLayout(Context context, AttributeSet attrs, int defStyleAttr) {
+    public ImageLayout(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
         init(context);
     }
@@ -44,11 +45,12 @@ public class InboxLayout extends ViewGroup {
         display.getSize(deviceDisplay);
         deviceWidth = deviceDisplay.x;
         deviceHeight = deviceDisplay.y;
+        //
+        inflate(getContext(), R.layout.image_layout, this);
 
-        inflate(getContext(), R.layout.inbox_layout, this);
 
         TitleText = findViewById(R.id.TitleText);
-        SearchText = findViewById(R.id.SearchText);
+        SearchText = findViewById(R.id.FilterButton);
         SearchInput = findViewById(R.id.SearchInput);
 
         SearchInput.setVisibility(View.GONE);
@@ -86,8 +88,8 @@ public class InboxLayout extends ViewGroup {
         for(int i = 0; i < count; i++){
             View temp = getChildAt(i);
 
-            if(temp instanceof InboxViewItem){
-                InboxViewItem child = (InboxViewItem) getChildAt(i);
+            if(temp instanceof ImageLayoutItem){
+                ImageLayoutItem child = (ImageLayoutItem) getChildAt(i);
 
                 if(!child.getTitle().toLowerCase().contains(seachText.toLowerCase())){
                     child.setVisibility(View.GONE);
@@ -95,6 +97,8 @@ public class InboxLayout extends ViewGroup {
                 else{
                     child.setVisibility(View.VISIBLE);
                 }
+
+                child.getMeasuredState();
             }
         }
 
@@ -117,6 +121,7 @@ public class InboxLayout extends ViewGroup {
     protected void onLayout(boolean changed, int l, int t, int r, int b) {
         final int count = getChildCount();
         int curWidth, curHeight, curLeft, curTop, maxHeight;
+
         //get the available size of child view
         final int childLeft = this.getPaddingLeft();
         final int childTop = this.getPaddingTop();
@@ -124,13 +129,17 @@ public class InboxLayout extends ViewGroup {
         final int childBottom = this.getMeasuredHeight() - this.getPaddingBottom();
         final int childWidth = childRight - childLeft;
         final int childHeight = childBottom - childTop;
+
         maxHeight = 0;
         curLeft = childLeft;
         curTop = childTop;
+
         for (int i = 0; i < count; i++) {
             View child = getChildAt(i);
+
             if (child.getVisibility() == GONE)
                 return;
+
             //Get the maximum size of the child
             child.measure(MeasureSpec.makeMeasureSpec(childWidth, MeasureSpec.AT_MOST), MeasureSpec.makeMeasureSpec(childHeight, MeasureSpec.AT_MOST));
             curWidth = child.getMeasuredWidth();
@@ -150,6 +159,50 @@ public class InboxLayout extends ViewGroup {
         }
     }
 
+
+
+    @Override
+    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+        int count = getChildCount();
+        // Measurement will ultimately be computing these values.
+        int maxHeight = 0;
+        int maxWidth = 0;
+        int childState = 0;
+        int mLeftWidth = 0;
+        int rowCount = 0;
+
+        // Iterate through all children, measuring them and computing our dimensions
+        // from their size.
+        for (int i = 0; i < count; i++) {
+            final View child = getChildAt(i);
+
+            if (child.getVisibility() == GONE)
+                continue;
+
+            // Measure the child.
+            measureChild(child, widthMeasureSpec, heightMeasureSpec);
+            maxWidth += Math.max(maxWidth, child.getMeasuredWidth());
+            mLeftWidth += child.getMeasuredWidth();
+
+            if ((mLeftWidth / deviceWidth) > rowCount) {
+                maxHeight += child.getMeasuredHeight();
+                rowCount++;
+            } else {
+                maxHeight = Math.max(maxHeight, child.getMeasuredHeight());
+            }
+            childState = combineMeasuredStates(childState, child.getMeasuredState());
+        }
+
+        // Check against our minimum height and width
+        maxHeight = Math.max(maxHeight, getSuggestedMinimumHeight());
+        maxWidth = Math.max(maxWidth, getSuggestedMinimumWidth());
+
+        // Report our final dimensions.
+        setMeasuredDimension(resolveSizeAndState(maxWidth, widthMeasureSpec, childState),
+                resolveSizeAndState(maxHeight, heightMeasureSpec, childState << MEASURED_HEIGHT_STATE_SHIFT));
+    }
+
+    /*
     @Override
     protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
         int count = getChildCount();
@@ -173,8 +226,6 @@ public class InboxLayout extends ViewGroup {
         maxWidth = Math.max(maxWidth, getSuggestedMinimumWidth());
         // Report our final dimensions.
         setMeasuredDimension(resolveSizeAndState(maxWidth, widthMeasureSpec, childState),
-        resolveSizeAndState(maxHeight, heightMeasureSpec, childState << MEASURED_HEIGHT_STATE_SHIFT));
-    }
-
-
+                resolveSizeAndState(maxHeight, heightMeasureSpec, childState << MEASURED_HEIGHT_STATE_SHIFT));
+    }*/
 }
