+++ cmykui/src/main/java/hr/foi/air/cmykui/layout/ImageLayoutItem.java
@@ -1,10 +1,7 @@
-package com.cmykui.framework.cmykui.layout;
+package hr.foi.air.cmykui.layout;
 
-import android.app.AlertDialog;
 import android.content.Context;
 import android.graphics.Bitmap;
-import android.graphics.Canvas;
-import android.graphics.drawable.BitmapDrawable;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.ImageView;
@@ -12,8 +9,8 @@ import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
-import com.cmykui.framework.cmykui.R;
-import com.cmykui.framework.cmykui.base.OnClick;
+import hr.foi.air.cmykui.R;
+import hr.foi.air.cmykui.base.OnClick;
 
 public class ImageLayoutItem extends RelativeLayout {
