+++ cmykui/src/main/java/hr/foi/air/cmykui/layout/ImageLayoutItemMenuItem.java
@@ -1,17 +1,16 @@
-package com.cmykui.framework.cmykui.layout;
+package hr.foi.air.cmykui.layout;
 
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.graphics.Bitmap;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
-import com.cmykui.framework.cmykui.R;
-import com.cmykui.framework.cmykui.base.OnClick;
+import hr.foi.air.cmykui.R;
+import hr.foi.air.cmykui.base.OnClick;
 
 public class ImageLayoutItemMenuItem extends RelativeLayout {
