+++ cmykui/src/main/res/layout/barchart_graph.xml
@@ -3,7 +3,7 @@
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.cmykui.framework.cmykui.graph.BarChart
+    <hr.foi.air.cmykui.graph.BarChart
         android:id="@+id/barchart"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
