+++ cmykui/src/main/res/layout/linechart_graph.xml
@@ -3,7 +3,7 @@
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.cmykui.framework.cmykui.graph.LineChart
+    <hr.foi.air.cmykui.graph.LineChart
         android:id="@+id/linechart"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
