+++ cmykui/src/main/res/layout/piechart_graph.xml
@@ -3,7 +3,7 @@
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.cmykui.framework.cmykui.graph.PieChart
+    <hr.foi.air.cmykui.graph.PieChart
         android:id="@+id/piechart"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
